renamed frontend to server since it is doing everything now anyway
