--- src/frontend/public/javascript/reconnect.js
@@ -1,4 +1,0 @@
-var saveGameinCookie = parseJwt(getCookie('chess-token'));
-if(saveGameinCookie != "" && saveGameinCookie != null && window.location.href.match(/play\/[a-zA-Z0-9]+/i) == null) {
-    window.location.replace('/play/' + saveGameinCookie.gameid);
-};
