--- src/frontend/views/error.ejs
@@ -1,12 +1,0 @@
-<script>
-    setCookie('chess-token', '');
-</script>
-<div class="">
-    <p class="">Oops, we encountered an error!</p>
-    <p class=""><%= error %></p>
-</div>
-<div class="w3-container w3-content w3-center w3-padding-64" style="max-width:800px" id="band">
-	<h2 class="w3-wide error">ERROR</h2>
-	<p class="w3-opacity"><i>We sadly encountered an error</i></p>
-    <p class=""><%= error %></p>
-</div>
