--- src/frontend/views/partials/footer.ejs
@@ -1,3 +1,0 @@
-<footer class="w3-container w3-padding-64 w3-center w3-opacity w3-light-grey w3-xlarge">
-  <p class="w3-medium">Copyright Niklas Oberhuber</p>
-</footer>
