--- src/frontend/views/partials/header.ejs
@@ -1,12 +1,0 @@
-<!-- Navbar -->
-<div class="w3-top">
-    <div class="w3-bar w3-black w3-card">
-        <a class="w3-bar-item w3-button w3-padding-large w3-hide-medium w3-hide-large w3-right" href="javascript:void(0)" onclick="myFunction()" title="Toggle Navigation Menu"><i class="fa fa-bars"></i></a>
-        <a href="https://chess.nikober.xyz" class="w3-bar-item w3-button w3-padding-large">CHESS.NIKOBER.XYZ</a>
-        <a href="#rules" class="w3-bar-item w3-button w3-padding-large w3-hide-small">RULES</a>
-    </div>
-</div>
-<!-- Navbar on small screens (remove the onclick attribute if you want the navbar to always show on top of the content when clicking on the links) -->
-<div id="navDemo" class="w3-bar-block w3-black w3-hide w3-hide-large w3-hide-medium w3-top" style="margin-top:46px">
-    <a href="#rules" class="w3-bar-item w3-button w3-padding-large" onclick="myFunction()">RULES</a>
-</div>
