--- src/frontend/views/play.ejs
@@ -1,24 +1,0 @@
-<script src="https://cdnjs.cloudflare.com/ajax/libs/socket.io/2.1.1/socket.io.js"></script>
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
-<script src="https://cdnjs.cloudflare.com/ajax/libs/chess.js/0.10.2/chess.js"></script>
-<script defer src="https://koblenski.github.io/javascript/chessboardjs-0.3.0/js/chessboard-0.3.0.min.js" type="text/javascript"></script>
-<link href="https://koblenski.github.io/javascript/chessboardjs-0.3.0/css/chessboard-0.3.0.min.css" rel="stylesheet"></link>
-
-<script defer src="/javascript/game.js"></script>
-<script defer src="/javascript/changeBoard.js"></script>
-
-<div class="w3-container w3-content w3-center w3-padding-64" style="max-width:800px" id="band">
-    <h2 id="roomIdcontainer" class="w3-wide"><p id="roomId"><%= gameid %></p></h2>
-    <p class="w3-opacity"><i><p id="gameStatus">Waiting for a second player.</p></i></p>
-    <br>
-
-    <div id="board" style="width: 100%; margin: auto"></div>
-
-    <br>
-    <button id="takebackButton" class="warning" onclick="offerTakeback(this)" disabled>Offer takeback.</button>
-    <div id="takeback">
-
-    </div>
-    <br>
-    <button class="danger" onclick="leaveMatch()">Leave Match!</button>
-</div>
