+++ src/server/public/javascript/JWTDecoder.js
