+++ src/server/public/javascript/reconnect.js
@@ -1,0 +1,4 @@
+var saveGameinCookie = parseJwt(getCookie('chess-token'));
+if(saveGameinCookie != "" && saveGameinCookie != null && window.location.href.match(/play\/[a-zA-Z0-9]+/i) == null) {
+    window.location.replace('/play/' + saveGameinCookie.gameid);
+};
