+++ src/server/public/stylesheet/main.css
@@ -1,0 +1,37 @@
+.danger {
+    color: #750000;
+}
+.warning {
+    color: #8a6500;
+}
+button {
+    padding: 14px 40px;
+    font-size: 16px;
+    background-color: #e7e7e7;
+    border: none;
+    border-radius: 5px;
+    text-align: center;
+    text-decoration: none;
+    margin: 10px;
+    cursor: pointer;
+}
+
+button:disabled {
+    padding: 14px 40px;
+    font-size: 16px;
+    background-color: #747474;
+    border: none;
+    border-radius: 5px;
+    text-align: center;
+    text-decoration: none;
+    margin: 10px;
+    color: #7c7c7c;
+    cursor: pointer;
+}
+
+input {
+    border: none;
+    border-bottom: 2px solid #747474;
+    background-color: transparent;
+    background: none;
+}
