+++ src/server/views/index.ejs
@@ -1,0 +1,18 @@
+<div class="w3-container w3-content w3-center w3-padding-64" style="max-width:800px" id="band">
+	<h2 class="w3-wide">PLAY A GAME</h2>
+	<p class="w3-opacity"><i>Play chess with your friends</i></p>
+	<p class="w3-justify">
+		Share the generated codes to play with your friend, or send them the link.
+	</p>
+	<div class="w3-row w3-padding-32">
+        <input type="text" id="gameidInput">
+		<br>
+		<br>
+		<Button type="submit" onclick="enterGamewithCode()" class="primary-button">Enter this game!</Button>
+	</div>
+	<div class="w3-row w3-padding-32">
+		<form action="/createGame" method="post">
+			<Button type="submit" class="primary-button">Create a game!</Button>
+	  	</form>
+	</div>
+</div>
